#84 and #119: simplify suggestion logic, second step - uint for counters

+++ internal/model/github/contribution/heatmap_test.go
@@ -71,7 +71,7 @@ func TestBuildHeatMap(t *testing.T) {
 
 	chm := BuildHeatMap(doc)
 	ts := xtime.UTC().Year(2019).Month(time.November).Day(13).Time()
-	assert.Equal(t, 3, chm.Count(ts))                   // 2019-11-13
-	assert.Equal(t, 2, chm.Count(ts.AddDate(0, 1, 0)))  // 2019-12-13
-	assert.Equal(t, 3, chm.Count(ts.AddDate(0, 1, 14))) // 2019-12-27
+	assert.Equal(t, uint(3), chm.Count(ts))                   // 2019-11-13
+	assert.Equal(t, uint(2), chm.Count(ts.AddDate(0, 1, 0)))  // 2019-12-13
+	assert.Equal(t, uint(3), chm.Count(ts.AddDate(0, 1, 14))) // 2019-12-27
 }
